+++ ImageFanReloaded.Core/ImageHandling/IImageFileFactory.cs
@@ -2,5 +2,5 @@ namespace ImageFanReloaded.Core.ImageHandling;
 
 public interface IImageFileFactory
 {
-    IImageFile GetImageFile(string fileName, string filePath, int sizeOnDiscInKilobytes);
+    IImageFile GetImageFile(string fileName, string filePath, decimal sizeOnDiscInKilobytes);
 }
